Dump long yaml fields in block format
The "output" and "traceback" attributes of MigrationSession.state tend
to be long and span multiple lines. By default, yaml.safe_dump formats
these as inline string with newline characters (e.g. "some text\nsome more text")
This is ugly and makes debugging difficult.

This change uses a subclass of yaml.SafeDumper to set strings wrapped
with the Output class to the block style, which is considerably easier
to read. It looks something like this:

    key: |
      some text
      some more text