Handle exception not tied to specific migration
In certain cases, all migrations can succeed but an error will still be
raised. In those cases, we still want to parse the output and save it in
the state.

This change also introduces the state['success'] field, which is a bool
that indicates the overall success of the step. This required renaming a
couple of the existing fields. The extra field should prevent clients
from having to check both state['traceback'] and
state['failed_migration'] to check for failure.